+++ radar-commons-android-qr/src/main/java/org/radarcns/android/auth/QrLoginManager.java
