+++ radar-commons-android-qr/src/main/res/values/strings.xml
@@ -1,0 +1,3 @@
+<resources>
+    <string name="app_name">radar-commons-android-oauth2</string>
+</resources>
