+++ settings.gradle
@@ -1,0 +1,2 @@
+include ':radar-commons-android-oauth2'
+include ':radar-commons-android-qr'
